Struct re_space_view_spatial::contexts::transform_context::TransformContext

pub struct TransformContext {
    space_origin: EntityPath,
    transform_per_entity: IntMap<EntityPath, TransformInfo>,
    unreachable_descendants: Vec<(EntityPath, UnreachableTransformReason)>,
    first_unreachable_parent: Option<(EntityPath, UnreachableTransformReason)>,
}

Provides transforms from an entity to a chosen reference space for all elements in the scene for the currently selected time & timeline.

The renderer then uses this reference space as its world space, making world and reference space equivalent for a given space view.

Should be recomputed every frame.

Fields
space_origin: EntityPath

All transforms provided are relative to this reference path.

transform_per_entity: IntMap<EntityPath, TransformInfo>

All reachable entities.

unreachable_descendants: Vec<(EntityPath, UnreachableTransformReason)>

All unreachable descendant paths of reference_path.

first_unreachable_parent: Option<(EntityPath, UnreachableTransformReason)>

The first parent of reference_path that is no longer reachable.

Implementations

impl TransformContext

fn gather_descendants_transforms( &mut self, tree: &EntityTree, data_store: &DataStore, query: &LatestAtQuery, entity_properties: &EntityPropertyMap, reference_from_entity: Affine3A, encountered_pinhole: &Option<EntityPath> )

pub fn reference_path(&self) -> &EntityPath

pub fn reference_from_entity(&self, ent_path: &EntityPath) -> Option<Affine3A>

Retrieves the transform of on entity from its local system to the space of the reference.

Returns None if the path is not reachable.

pub fn reference_from_entity_ignoring_pinhole( &self, ent_path: &EntityPath, store: &DataStore, query: &LatestAtQuery ) -> Option<Affine3A>

Like Self::reference_from_entity, but if ent_path has a pinhole camera, it won’t affect the transform.

Normally, the transform we compute for an entity with a pinhole transform places all objects in front (defined by view coordinates) of the camera with a given image plane distance. In some cases like drawing the lines for a frustum or arrows for the 3D transform, this is not the desired transformation. Returns None if the path is not reachable.

TODO(#2663, #1025): Going forward we should have separate transform hierarchies for 2D (i.e. projected) and 3D, which would remove the need for this.

pub fn parent_pinhole(&self, ent_path: &EntityPath) -> Option<&EntityPath>

Retrieves the ancestor (or self) pinhole under which this entity sits.

None indicates either that the entity does not exist in this hierarchy or that this entity is under the eye camera with no Pinhole camera in-between. Some indicates that the entity is under a pinhole camera at the given entity path that is not at the root of the space view.