Module re_space_view_spatial::parts

Responsible for populating SceneSpatialPrimitives and SceneSpatialUiData

Re-exports

pub use cameras::CamerasPart;

pub use images::ImagesPart;

pub use images::ViewerImage;

pub use spatial_view_part::SpatialViewPartData;

pub use transform3d_arrows::add_axis_arrows;

pub use transform3d_arrows::Transform3DArrowsPart;

Modules

arrows3d 🔒

assets3d 🔒

boxes2d 🔒

boxes3d 🔒

cameras 🔒

entity_iterator 🔒

images 🔒

lines2d 🔒

lines3d 🔒

meshes 🔒

points2d 🔒

points3d 🔒

spatial_view_part 🔒

transform3d_arrows 🔒

Structs

UiLabel

Enums

UiLabelTarget

Constants

SIZE_BOOST_IN_POINTS_FOR_LINE_OUTLINES

SIZE_BOOST_IN_POINTS_FOR_POINT_OUTLINES

Functions

calculate_bounding_box

collect_ui_labels

image_view_coordinates

Returns the view coordinates used for 2D (image) views.

load_keypoint_connections

picking_id_from_instance_key

process_annotations 🔒

Resolves all annotations for the given entity view.

process_annotations_and_keypoints 🔒

Resolves all annotations and keypoints for the given entity view.

process_colors

Process Color components using annotations and default colors.

process_labels

Process Text components using annotations.

process_radii

Process re_types::components::Radius components to re_renderer::Size using auto size where no radius is specified.

register_parts

Type Definitions

Keypoints

Collection of keypoints for annotation context.