Module re_space_view_spatial::visualizers

Responsible for populating SceneSpatialPrimitives and SceneSpatialUiData

Re-exports

pub use cameras::CamerasVisualizer;

pub use images::ImageVisualizer;

pub use images::ViewerImage;

pub use spatial_view_visualizer::SpatialViewVisualizerData;

pub use transform3d_arrows::add_axis_arrows;

pub use transform3d_arrows::Transform3DArrowsVisualizer;

Modules

arrows2d 🔒

arrows3d 🔒

assets3d 🔒

boxes2d 🔒

boxes3d 🔒

cameras 🔒

entity_iterator 🔒

images 🔒

lines2d 🔒

lines3d 🔒

meshes 🔒

points2d 🔒

points3d 🔒

results_ext 🔒

spatial_view_visualizer 🔒

transform3d_arrows 🔒

Structs

UiLabel

Enums

UiLabelTarget

Constants

SIZE_BOOST_IN_POINTS_FOR_LINE_OUTLINES

SIZE_BOOST_IN_POINTS_FOR_POINT_OUTLINES

Functions

collect_ui_labels

filter_visualizable_2d_entities 🔒

filter_visualizable_3d_entities 🔒

image_view_coordinates

Returns the view coordinates used for 2D (image) views.

load_keypoint_connections

process_annotation_and_keypoint_slices 🔒

Resolves all annotations and keypoints for the given entity view.

process_color_slice

Process Color components using annotations and default colors.

process_radius 🔒

process_radius_slice

Process re_types::components::Radius components to re_renderer::Size using auto size where no radius is specified.

register_2d_spatial_visualizers

register_3d_spatial_visualizers

Type Aliases

Keypoints

Collection of keypoints for annotation context.