Function re_space_view_spatial::visualizers::entity_iterator::process_archetype

pub fn process_archetype<System: IdentifiedViewSystem, A, F>(
    ctx: &ViewerContext<'_>,
    query: &ViewQuery<'_>,
    view_ctx: &ViewContextCollection,
    default_depth_offset: DepthOffset,
    fun: F
) -> Result<(), SpaceViewSystemExecutionError>
where
    A: Archetype,
    F: FnMut(&ViewerContext<'_>, &EntityPath, &EntityProperties, &SpatialSceneEntityContext<'_>, &Results) -> Result<(), SpaceViewSystemExecutionError>,

Iterates through all entity views for a given archetype.

The callback passed in gets passed along an SpatialSceneEntityContext which contains various useful information about an entity in the context of the current scene.