Trait re_space_view_spatial::visualizers::results_ext::RangeResultsExt

pub trait RangeResultsExt {
    // Required methods
    fn get_dense<'a, C: Component>(
        &'a self,
        resolver: &PromiseResolver
    ) -> Option<Result<RangeData<'a, C>>>;
    fn get_or_empty_dense<'a, C: Component>(
        &'a self,
        resolver: &PromiseResolver
    ) -> Result<RangeData<'a, C>>;
}

Extension traits to abstract query result handling for all spatial space views.

Also turns all results into range results, so that views only have to worry about the ranged case.

Required Methods

fn get_dense<'a, C: Component>( &'a self, resolver: &PromiseResolver ) -> Option<Result<RangeData<'a, C>>>

fn get_or_empty_dense<'a, C: Component>( &'a self, resolver: &PromiseResolver ) -> Result<RangeData<'a, C>>

Object Safety

This trait is not object safe.