Function re_space_view_spatial::visualizers::utilities::labels::show_labels_fallback

pub fn show_labels_fallback<C: Component>(ctx: &QueryContext<'_>) -> ShowLabels

Given a visualizer’s query context, compute its ShowLabels fallback value (used when neither the logged data nor the blueprint provides a value).

Assumes that the visualizer reads the Text component for components. The type parameter C must be the component type that defines the number of instances in the batch.

This function is normally used to implement the ComponentFallbackProvider that will be used in a LabeledBatch.