Module re_space_view_spatial::visualizers::utilities::proc_mesh_vis

Structs

ProcMeshBatch

A batch of instances to draw. This struct is just arguments to ProcMeshDrawableBuilder::add_batch().

ProcMeshDrawableBuilder

To be used within the scope of a single VisualizerSystem::execute() call when the visualizer wishes to draw batches of ProcMeshKey meshes.